XCCL: fix review comments
